Move dict tests to separate module. Add more tests.
