+++ tests/conftest.py
@@ -2,19 +2,11 @@ import pytest
 import redis
 
 
-def _get_client(cls, request=None, **kwargs):
-    params = {'host': 'localhost', 'port': 6379, 'db': 9}
-    params.update(kwargs)
-    client = cls(**params)
-    client.flushdb()
-    if request:
-        def teardown():
-            client.flushdb()
-            client.connection_pool.disconnect()
-        request.addfinalizer(teardown)
-    return client
-
-
 @pytest.fixture()
-def r(request, **kwargs):
-    return _get_client(redis.Redis, request, **kwargs)
+def r():
+    """Redis client."""
+    client = redis.Redis(host='localhost', port=6379, db=9)
+    client.flushdb()
+    yield client
+    client.flushdb()
+    client.connection_pool.disconnect()
